fix: mergeability — consolidate projection files for pkg-file-count limit
Fold work-move API delegate and dashboard move-operation builder into
existing package files to stay within the 15-file per-package cap.
