+++ pkg/api/workstation_request_projection.go
@@ -2,6 +2,7 @@ package api
 
 import (
 	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
+	"github.com/portpowered/infinite-you/pkg/api/moveprojection"
 	"github.com/portpowered/infinite-you/pkg/api/workstationprojection"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 )
@@ -12,4 +13,12 @@ func BuildFactoryWorldWorkstationRequestProjectionSlice(
 	state interfaces.FactoryWorldState,
 ) factoryapi.FactoryWorldWorkstationRequestProjectionSlice {
 	return workstationprojection.BuildFactoryWorldWorkstationRequestProjectionSlice(state)
+}
+
+// BuildFactoryWorldWorkMoveOperationProjectionSlice delegates to the
+// moveprojection subpackage while preserving the historical pkg/api entrypoint.
+func BuildFactoryWorldWorkMoveOperationProjectionSlice(
+	state interfaces.FactoryWorldState,
+) factoryapi.FactoryWorldWorkMoveOperationProjectionSlice {
+	return moveprojection.BuildFactoryWorldWorkMoveOperationProjectionSlice(state)
 }
