+++ pkg/factory/projections/simple_dashboard_projection.go
@@ -114,3 +114,28 @@ func appendCustomerOutputPlaceIDs(workstation interfaces.FactoryWorkstation) []s
 	outputPlaceIDs = append(outputPlaceIDs, workstation.FailurePlaceIDs...)
 	return outputPlaceIDs
 }
+
+func buildWorkMoveOperationsByWorkID(
+	state interfaces.FactoryWorldState,
+) map[string][]interfaces.FactoryWorldWorkStateChangeRecord {
+	if len(state.WorkStateChangesByWorkID) == 0 {
+		return nil
+	}
+
+	operationsByWorkID := make(
+		map[string][]interfaces.FactoryWorldWorkStateChangeRecord,
+		len(state.WorkStateChangesByWorkID),
+	)
+	for workID, records := range state.WorkStateChangesByWorkID {
+		if len(records) == 0 {
+			continue
+		}
+		cloned := make([]interfaces.FactoryWorldWorkStateChangeRecord, len(records))
+		copy(cloned, records)
+		operationsByWorkID[workID] = cloned
+	}
+	if len(operationsByWorkID) == 0 {
+		return nil
+	}
+	return operationsByWorkID
+}
